Refactor ProjectService
