+++ client/src/utils.ts
@@ -1,12 +1,8 @@
 import { mkdtemp, readFileSync, unlink } from "fs";
 import { tmpdir } from "os";
 import { dirname, join } from "path";
-import { commands, extensions, Uri, window, workspace } from "vscode";
-import {
-  CMD_GENERATE_EXTENSION, EXTENSION_GENERATOR_TARGET, VS_CMD_INSTALL_EXTENSION,
-  VS_CMD_UNINSTALL_EXTENSION,
-} from "./constants";
-import { ICommand, ITextXExtensionInstall, ITextXExtensionUninstall } from "./interfaces";
+import { extensions, window, workspace } from "vscode";
+import { ICommand } from "./interfaces";
 import { TokenColors } from "./types";
 
 export function getCommands(command: string): ICommand {
@@ -58,25 +54,6 @@ export function getTokenColorsForTheme(themeName: string): TokenColors {
   return tokenColors;
 }
 
-export function generateAndInstallExtension(projectName: string): Promise<ITextXExtensionInstall> {
-  return new Promise(async (resolve) => {
-    mkdtempWrapper(async (folder) => {
-      const extensionPath = await commands.executeCommand<string>(
-        CMD_GENERATE_EXTENSION.external, projectName, EXTENSION_GENERATOR_TARGET, folder,
-      );
-
-      extensions.onDidChange((_) => {
-        const extensionName = extensionPath.split("\\").pop().split("/").pop().split(".")[0];
-        const extension = extensions.getExtension(textxExtensionName(extensionName));
-        const isActive = extension === undefined ? false : extension.isActive;
-        resolve({extension, isActive, isInstalled: extension !== undefined });
-      });
-
-      await commands.executeCommand(VS_CMD_INSTALL_EXTENSION, Uri.file(extensionPath));
-    });
-  });
-}
-
 export function mkdtempWrapper(callback: (folder: string) => Promise<void>): void {
   mkdtemp(join(tmpdir(), "textx-"), async (err, folder) => {
     if (err) {
@@ -84,27 +61,6 @@ export function mkdtempWrapper(callback: (folder: string) => Promise<void>): voi
     }
     await callback(folder);
     unlink(folder, (unlinkErr) => unlinkErr );
-  });
-}
-
-export function uninstallExtension(projectName: string): Promise<ITextXExtensionUninstall> {
-  return new Promise(async (resolve) => {
-    const extensionName = textxExtensionName(projectName);
-    let extension = extensions.getExtension(extensionName);
-    const isActive = extension === undefined ? false : extension.isActive;
-
-    extensions.onDidChange((_) => {
-      // If extension is NOT active
-      extension = extensions.getExtension(extensionName);
-      resolve({ isActive, isUninstalled: extension === undefined });
-    });
-
-    try {
-      await commands.executeCommand(VS_CMD_UNINSTALL_EXTENSION, extensionName);
-      resolve({ isActive, isUninstalled: true });
-    } catch (_) {
-      resolve({ isActive, isUninstalled: false});
-    }
   });
 }
